fix: Empty cells exception (#228)
This commit let's the model know if cells are empty. Previously, we would throw an exception so the user could see that cells are empty, under the assumption that the user will never provide empty cells on purpose. That was perhaps naive. #218 shows a real-life example where it is better to just let the model know that all cells were empty, so it can make appropriate an response.